+++ src/nice_int/mod.rs
@@ -76,7 +76,7 @@ impl<const S: usize> Deref for NiceWrapper<S> {
 impl<const S: usize> fmt::Display for NiceWrapper<S> {
 	#[inline]
 	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		f.write_str(self.as_str())
+		f.pad(self.as_str())
 	}
 }